+++ frontend/.env.example
@@ -1,0 +1,1 @@
+VITE_BACKEND_URL=
