+++ frontend/src/components/UserTable.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import { useFetchUser } from "../config/fetch";
-import UsersPagination from "@/components/pagination/users.pagination";
-import UserCreateModal from "@/components/modal/user.create.modal";
-import UserEditModal from "@/components/modal/user.edit.modal";
-import UserDeleteModal from "@/components/modal/user.delete.modal";
+import UsersPagination from "@/components/pagination/UserPagnation";
+import UserCreateModal from "@/components/modal/UserCreate";
+import UserEditModal from "@/components/modal/UserEdit";
+import UserDeleteModal from "@/components/modal/UserDelete";
 // import { Popover } from "@/components/ui/popover";
 function UsersTable() {
     interface IUser {
