+++ frontend/src/main.tsx
@@ -6,12 +6,12 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 // import ProtectedRoute from '@/components/auth/index.tsx'
 import { Toaster } from './components/ui/sonner.tsx'
 // import { SocketProvider } from './components/context/socket.context.tsx'
-import HomePage from './pages/home.tsx'
+import HomePage from '@/pages/home.tsx'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import Layout from './layout.tsx'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-import LoginPage from './pages/login.tsx'
-import RegisterPage from './pages/register.tsx'
+import LoginPage from '@/pages/login.tsx'
+import RegisterPage from '@/pages/register.tsx'
 
 const router = createBrowserRouter([
   {
