+++ frontend/src/types/api.d.ts
@@ -1,0 +1,16 @@
+interface IBackendRes<T> {
+  error?: string | string[];
+  message: string;
+  statusCode: number | string;
+  data?: T;
+}
+
+interface IModelPaginate<T> {
+  meta: {
+    current: number;
+    pageSize: number;
+    pages: number;
+    total: number;
+  };
+  result: T[];
+}
